Move admin/site to src/Module
- Replace default theme settings modal with simple link to admin/themes
