+++ view/templates/admin/site.tpl
@@ -1,44 +1,3 @@
-<script>
-	$(function(){
-
-		$("#cnftheme").click(function(){
-			$.colorbox({
-				width: 800,
-				height: '90%',
-				/*onOpen: function(){
-					var theme = $("#id_theme :selected").val();
-					$("#cnftheme").attr('href',"{{$baseurl}}/admin/themes/"+theme);
-				},*/
-				iframe: true,
-				href: "{{$baseurl}}/admin/themes/" + $("#id_theme :selected").val() + "?mode=minimal",
-				onComplete: function(){
-					$("div#fancybox-content form").submit(function(e){
-						var url = $(this).attr('action');
-						// can't get .serialize() to work...
-						var data={};
-						$(this).find("input").each(function(){
-							data[$(this).attr('name')] = $(this).val();
-						});
-						$(this).find("select").each(function(){
-							data[$(this).attr('name')] = $(this).children(":selected").val();
-						});
-						console.log(":)", url, data);
-
-						$.post(url, data, function(data) {
-							if(timer) clearTimeout(timer);
-							NavUpdate();
-							$.colorbox.close();
-						})
-
-						return false;
-					});
-
-				}
-			});
-			return false;
-		});
-	});
-</script>
 <div id='adminpage'>
 	<h1>{{$title}} - {{$page}}</h1>
